requirements.txt
pynmea
pyserial